Initial Import of original RAM code
The original code used to be downloadable from
ftp://ftp.ccs.nrl.navy.mil/pub/ram/RAM
but this is no longer the case.

Unfortunately there was no LICENSE attached to the code, but the code is
used in many scientific publications, so one can assume it is in the
public domain.